src/middleware/waf.go
package middleware

import (
	"net/http"
)

// WAFMiddleware is a middleware that performs web application firewall checks
func WAFMiddleware(next http.Handler) http.Handler {
	middlewareFuncs := []func(http.Handler) http.Handler{
		InputValidationMiddleware,
		SanitizationMiddleware,
		RateLimitingMiddleware,
		AuthenticationMiddleware,
		AuthorizationMiddleware,
		APICallMiddleware,
	}

	// Chain all middleware functions together
	for i := len(middlewareFuncs) - 1; i >= 0; i-- {
		next = middlewareFuncs[i](next)
	}
	return next
}

// InputValidationMiddleware is a middleware that performs input validation on incoming requests
func InputValidationMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Example: Perform input validation on request parameters, headers, and body
		// ...
		next.ServeHTTP(w, r)
	})
}

// SanitizationMiddleware is a middleware that performs sanitization on incoming requests
func SanitizationMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Example: Sanitize request parameters, headers, and body to remove potentially harmful characters or scripts
		// ...
		next.ServeHTTP(w, r)
	})
}

// RateLimitingMiddleware is a middleware that implements rate limiting to prevent abuse or DoS attacks
func RateLimitingMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Example: Implement rate limiting to restrict the number of requests per IP address or user
		// ...
		next.ServeHTTP(w, r)
	})
}

// AuthenticationMiddleware is a middleware that performs user authentication
func AuthenticationMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Example: Authenticate users based on credentials or tokens
		// ...
		next.ServeHTTP(w, r)
	})
}

// AuthorizationMiddleware is a middleware that performs user authorization
func AuthorizationMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Example: Authorize users to access specific resources based on their roles or permissions
		// ...
		next.ServeHTTP(w, r)
	})
}
